Fix issue with undefined method `end_date`
* Was not failing in spec because methods in cucumber were part of
  global namespace
* Add helper methods to module, which results in failure
* Fix failure
* Boom
