+++ features/admin_views_open_auction.feature
@@ -6,13 +6,11 @@ Feature: Admin views open auction
     Given there is an open auction
     And I am an administrator
     And I sign in
-    When I visit the auction page
-    Then I should see the auction's title
-    And I should not see the bid form
-    And I should the open auction message for admins
-
     When I visit the admin auction page for that auction
-    Then I should the open auction message for admins
+    Then I should see the open auction message for admins
+
+    When I visit the auction page
+    Then I should see the open auction message for admins
 
   Scenario: There is an open auction for the other purchase card
     Given there is an open auction
@@ -20,4 +18,4 @@ Feature: Admin views open auction
     And I am an administrator
     And I sign in
     When I visit the admin auction page for that auction
-    Then I should the open auction message for admins
+    Then I should see the open auction message for admins
